+++ app/src/main/res/drawable/bottom_nav_bar_selector.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector
+    xmlns:android="http://schemas.android.com/apk/res/android" xmlns:app="http://schemas.android.com/apk/res-auto" android:constantSize="false">
+    <item
+        android:state_pressed="true"
+        android:state_checked="false"
+        android:color="@color/primary"
+        />
+    <item
+        android:state_pressed="false"
+        android:state_checked="false"
+        android:color="@color/primary_light"
+        />
+    <item
+        android:state_checked="true"
+        android:color="@color/primary_dark"
+        />
+</selector>
