--- app/src/main/res/menu/activity_main_drawer.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
-
-    <group
-        android:id="@+id/nav_group_general"
-        android:checkableBehavior="single">
-
-        <item
-            android:id="@+id/nav_main_page"
-            android:icon="@drawable/ic_menu_main"
-            android:title="@string/nav_main" />
-    </group>
-
-    <group
-        android:id="@+id/nav_group_settings"
-        android:checkableBehavior="single">
-
-        <item
-            android:id="@+id/nav_settings"
-            android:icon="@drawable/ic_menu_settings"
-            android:title="@string/nav_settings" />
-        <item
-            android:id="@+id/nav_about"
-            android:icon="@drawable/ic_menu_about"
-            android:title="@string/nav_about" />
-    </group>
-
-</menu>
