+++ app/src/main/res/values/strings.xml
@@ -26,5 +26,8 @@
     <string name="dollar_rate">Dollar rate:</string>
     <string name="usd" tools:ignore="MissingTranslation">USD</string>
     <string name="rub" tools:ignore="MissingTranslation">RUB</string>
+    <string name="title_home">Home</string>
+    <string name="title_dashboard">Dashboard</string>
+    <string name="title_notifications">Notifications</string>
 
 </resources>
